# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for bookmark management",
        "title": "Bookmark API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/genpass": {
            "get": {
                "description": "Generate a new password",
                "tags": [
                    "password"
                ],
                "summary": "Generate a new password",
                "responses": {
                    "200": {
                        "description": "12345678",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/links/shorten": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "link"
                ],
                "summary": "receive the url, return the code",
                "parameters": [
                    {
                        "description": "Shorten URL Input payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.shortenUrlRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/links/shorten/{code}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "link"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.shortenUrlRequest": {
            "type": "object",
            "required": [
                "exp",
                "url"
            ],
            "properties": {
                "exp": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        }
    }
}